youtube_processing/requirements.txt
yt-dlp 
openai-whisper
googletrans==4.0.0-rc1 
ffmpeg-python